src/kontrol/{utils,solc_to_k,state_record,prove}: move hex_string_to_int from solc_to_k to utils
